moves _overpass_json_from_file() and _OSMContentHandler() from graph to downloader
