+++ osmnx/graph.py
@@ -1,8 +1,5 @@
 """Graph creation functions."""
 
-import bz2
-import os
-import xml.sax
 from itertools import groupby
 
 import networkx as nx
@@ -494,7 +491,7 @@ def graph_from_xml(filepath, bidirectional=False, simplify=True, retain_all=Fals
     G : networkx.MultiDiGraph
     """
     # transmogrify file of OSM XML data into JSON
-    response_jsons = [_overpass_json_from_file(filepath)]
+    response_jsons = [downloader._overpass_json_from_file(filepath)]
 
     # create graph using this response JSON
     G = _create_graph(response_jsons, bidirectional=bidirectional, retain_all=retain_all)
@@ -505,34 +502,6 @@ def graph_from_xml(filepath, bidirectional=False, simplify=True, retain_all=Fals
 
     utils.log(f"graph_from_xml returned graph with {len(G)} nodes and {len(G.edges())} edges")
     return G
-
-
-def _overpass_json_from_file(filepath):
-    """
-    Read OSM XML from file and return Overpass-like JSON.
-
-    Parameters
-    ----------
-    filepath : string
-        path to file containing OSM XML data
-
-    Returns
-    -------
-    OSMContentHandler object
-    """
-
-    def _opener(filepath):
-        _, ext = os.path.splitext(filepath)
-        if ext == ".bz2":
-            return bz2.BZ2File(filepath)
-        else:
-            # assume an unrecognized file extension is just XML
-            return open(filepath, mode="rb")
-
-    with _opener(filepath) as file:
-        handler = _OSMContentHandler()
-        xml.sax.parse(file, handler)
-        return handler.object
 
 
 def _create_graph(response_jsons, retain_all=False, bidirectional=False):
@@ -767,57 +736,3 @@ def _add_paths(G, paths, bidirectional=False):
             _add_path(G, data, one_way=False)
 
     return G
-
-
-class _OSMContentHandler(xml.sax.handler.ContentHandler):
-    """
-    SAX content handler for OSM XML.
-
-    Used to build an Overpass-like response JSON object in self.object. For format
-    notes, see http://wiki.openstreetmap.org/wiki/OSM_XML#OSM_XML_file_format_notes
-    and http://overpass-api.de/output_formats.html#json
-    """
-
-    def __init__(self):
-        self._element = None
-        self.object = {"elements": []}
-
-    def startElement(self, name, attrs):
-        if name == "osm":
-            self.object.update({k: attrs[k] for k in attrs.keys() if k in {"version", "generator"}})
-
-        elif name in {"node", "way"}:
-            self._element = dict(type=name, tags={}, nodes=[], **attrs)
-            self._element.update({k: float(attrs[k]) for k in attrs.keys() if k in {"lat", "lon"}})
-            self._element.update(
-                {
-                    k: int(attrs[k])
-                    for k in attrs.keys()
-                    if k in {"id", "uid", "version", "changeset"}
-                }
-            )
-
-        elif name == "relation":
-            self._element = dict(type=name, tags={}, members=[], **attrs)
-            self._element.update(
-                {
-                    k: int(attrs[k])
-                    for k in attrs.keys()
-                    if k in {"id", "uid", "version", "changeset"}
-                }
-            )
-
-        elif name == "tag":
-            self._element["tags"].update({attrs["k"]: attrs["v"]})
-
-        elif name == "nd":
-            self._element["nodes"].append(int(attrs["ref"]))
-
-        elif name == "member":
-            self._element["members"].append(
-                {k: (int(attrs[k]) if k == "ref" else attrs[k]) for k in attrs.keys()}
-            )
-
-    def endElement(self, name):
-        if name in {"node", "way", "relation"}:
-            self.object["elements"].append(self._element)
